RD-5755 Part 4 move app/containers/status to app/components/status
