+++ app/components/status/SystemStatusIcon.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { connect } from 'react-redux';
+import type { ReduxState } from '../../reducers';
 
 import ClusterStatusIcon from '../shared/cluster/ClusterStatusIcon';
 import type { ClusterServiceStatus } from '../shared/cluster/types';
@@ -7,6 +9,19 @@ export interface SystemStatusIconProps {
     systemStatus?: ClusterServiceStatus;
 }
 
-export default function SystemStatusIcon({ systemStatus }: SystemStatusIconProps) {
+function SystemStatusIcon({ systemStatus }: SystemStatusIconProps) {
     return <ClusterStatusIcon status={systemStatus} />;
 }
+const mapStateToProps = (state: ReduxState) => {
+    const systemStatus = state.manager.clusterStatus.status;
+    return {
+        systemStatus,
+        maintenanceStatus: state.manager.maintenance
+    };
+};
+
+const mapDispatchToProps = () => {
+    return {};
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(SystemStatusIcon);
